requirements.txt
streamlit
python-dotenv
mcp
openai
python-docx
-e ./mcp-atlassian-main
chromadb
